requirements.txt
git+https://github.com/mccoderpy/discord.py-message-components.git@developer
python-dotenv~=0.21.0
discord~=2.1.0
p~=1.4.0
discord-py-slash-command~=4.2.1
nextcord~=2.3.2
aiohttp~=3.8.3
Pillow~=9.3.0
aiosqlite~=0.17.0